Fix memory test static analysis regressions
